+++ test/rubocop/cop/minitest/refute_includes_test.rb
@@ -66,11 +66,40 @@ class RefuteIncludesTest < Minitest::Test
     RUBY
   end
 
+  def test_registers_offense_when_using_refute_with_include_in_redundant_parentheses
+    assert_offense(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute((collection.include?(object)))
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer using `refute_includes(collection, object)` over `refute(collection.include?(object))`.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute_includes((collection, object))
+        end
+      end
+    RUBY
+  end
+
   def test_does_not_register_offense_when_using_refute_includes_method
     assert_no_offenses(<<~RUBY)
       class FooTest < Minitest::Test
         def test_do_something
           refute_includes(collection, object)
+        end
+      end
+    RUBY
+  end
+
+  def test_registers_offense_when_using_refute_with_non_include_method
+    assert_no_offenses(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute(collection.end_with?(string))
         end
       end
     RUBY